requirements_optimized.txt
# Hajimi King 优化版依赖
# 核心依赖
requests>=2.28.0
python-dotenv>=0.19.0
google-generativeai>=0.3.0

# 性能优化依赖
httpx>=0.24.0  # 异步HTTP客户端
aiofiles>=23.0.0  # 异步文件操作
tenacity>=8.2.0  # 重试机制
urllib3>=1.26.0  # 连接池支持

# Web监控依赖
flask>=2.3.0  # Web框架
flask-cors>=4.0.0  # CORS支持

# 数据库依赖
# SQLite3已内置Python，无需额外安装

# 通知系统依赖（可选）
# email-validator>=2.0.0  # 邮件验证
# slack-sdk>=3.21.0  # Slack通知
# discord.py>=2.3.0  # Discord通知

# 开发依赖（可选）
# pytest>=7.3.0  # 测试框架
# black>=23.0.0  # 代码格式化
# pylint>=2.17.0  # 代码检查
# mypy>=1.3.0  # 类型检查